refactor(domain): dedupe event-recording test helper, hoist relationship import
Extract the duplicated RecordingHandler + recording_publisher() test helper
(previously copy-pasted in coaching_session and coaching_session_goal test
modules) into a shared crate::test_support module, gated on
all(test, feature = "mock") so it compiles exactly when its consumers do.
Remove the now-orphaned async_trait/EventHandler/Arc/Mutex imports.

Also add a file-level use crate::coaching_relationships and drop the inline
crate:: qualification from publish_goals_linked and
find_by_id_with_coaching_relationship for consistency.

No behavior change. clippy -D warnings clean on default and mock features;
mock test suite green.
